Refactor VesselsLayer to typescript and fix bug when hiding vessels layer
