+++ frontend/src/features/map/layers/Vessel/VesselEstimatedPositionLayer.js
@@ -1,11 +1,11 @@
 import React, { useEffect, useRef } from 'react'
 import { useSelector } from 'react-redux'
 import VectorSource from 'ol/source/Vector'
-import { LayerProperties } from '../../../domain/entities/layers/constants'
-import { EstimatedPosition } from '../../../domain/entities/estimatedPosition'
-import { getVesselLastPositionVisibilityDates, Vessel, vesselIsShowed } from '../../../domain/entities/vessel/vessel'
+import { LayerProperties } from '../../../../domain/entities/layers/constants'
+import { EstimatedPosition } from '../../../../domain/entities/estimatedPosition'
+import { getVesselLastPositionVisibilityDates, Vessel, vesselIsShowed } from '../../../../domain/entities/vessel/vessel'
 import { Vector } from 'ol/layer'
-import { getEstimatedPositionStyle } from './styles/vesselEstimatedPosition.style'
+import { getEstimatedPositionStyle } from '../styles/vesselEstimatedPosition.style'
 
 const VesselEstimatedPositionLayer = ({ map }) => {
   const {
